requirements.txt
notebook
ipywidgets
jupyter
websocket-client==0.37.0
pillow
scipy
matplotlib
pyserial
opencv-python